Feature has been added to provide user with options of cities
API has been added to WeatherSearchBar

Weather search bar takes in City from user and provides a list of options for the user to pick that city that happens to be in different states/countries

This new API Take the users selections from the list, and passes the longitude and latitude coordinates to the API in WeatherAPI.js.

There it returns the weather information based on the lon and lat coordinates.